changed to_TGraph function placement
