+++ src/uproot/__init__.py
@@ -125,7 +125,7 @@ from uproot.writing import WritableBranch
 from uproot.writing import to_writable
 from uproot.writing import dask_write
 
-from uproot.writing.identify import to_TGraph
+from uproot.writing.interpret import to_TGraph
 
 import uproot.models.TObject
 import uproot.models.TString
